YAPC::EU 2018 Glasgow Update!

+++ Task/Count-in-octal/Nim/count-in-octal.nim
@@ -1,3 +1,3 @@
 import strutils
-for i in 0 .. <int64.high:
+for i in 0 ..< int.high:
   echo toOct(i, 16)
